HW-16: Create policies for posts and comments. Add gates for access to admin-panel and paid functionality

+++ app/Http/Controllers/Admin/PostController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\Post;
 use App\Models\Tag;
@@ -11,15 +12,17 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Validation\Rule;
 use Illuminate\View\View;
 
-class PostController
+class PostController extends Controller
 {
     /**
      * Display listing of posts.
      *
      * @return View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function index(): View
     {
+        $this->authorize('viewAny', Post::class);
         $posts = Post::with(['tags', 'category', 'user'])->paginate(10);
 
         return view('admin/post/index', compact('posts'));
@@ -30,11 +33,12 @@ class PostController
      *
      * @param int $id
      * @return View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(int $id): View
     {
         $post = Post::find($id);
-
+        $this->authorize('view', $post);
         return view('admin/post/show', compact('post'));
     }
 
@@ -42,9 +46,11 @@ class PostController
      * Display form for creating new post
      *
      * @return View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create(): View
     {
+        $this->authorize('create', Post::class);
         $post = new Post();
         $categories = Category::all();
         $tags = Tag::all();
@@ -57,9 +63,11 @@ class PostController
      *
      * @param Request $request
      * @return RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request): RedirectResponse
     {
+        $this->authorize('create', Post::class);
         $request->validate([
             'title' => [
                 'required',
@@ -91,10 +99,12 @@ class PostController
      *
      * @param int $id
      * @return View
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(int $id): View
     {
         $post = Post::find($id);
+        $this->authorize('update', $post);
         $categories = Category::all();
         $tags = Tag::all();
         return view('admin/post/form-edit', compact('post', 'categories', 'tags'));
@@ -105,11 +115,13 @@ class PostController
      *
      * @param Request $request
      * @return RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request): RedirectResponse
     {
         $id = $request->input('id');
         $post = Post::find($id);
+        $this->authorize('update', $post);
         $request->validate([
             'title' => [
                 'required',
@@ -141,12 +153,14 @@ class PostController
      *
      * @param int $id
      * @return RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(int $id): RedirectResponse
     {
         $post = Post::find($id);
+        $this->authorize('delete', $post);
         $post->tags()->detach();
         $post->delete();
-        return redirect()->route('admin.post');
+        return redirect()->back();
     }
 }
